pyproject.toml
[tool.poetry]
name = "cryoet"
version = "0.1.0"
description = "Work about competition CryoET Object Identification"
authors = ["henriupton99 <henriupton99@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
zarr = "^2.18.3"
numpy = "^2.1.3"
matplotlib = "^3.9.2"
plotly = "^5.24.1"
pyyaml = "^6.0.2"
pandas = "^2.2.3"
pytest = "^8.3.3"
kaleido = "^0.4.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
